requirements.txt
# Core ML
torch>=2.1.0
transformers>=4.37.0
einops>=0.7.0
jaxtyping>=0.2.25

# Mechanistic interpretability
sae-lens>=3.0.0
nnsight>=0.3.0

# Activation oracle (LoRA)
peft>=0.7.0

# CKA & alignment
scikit-learn>=1.3.0
scipy>=1.11.0
numpy>=1.24.0

# Datasets
datasets>=2.16.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Experiment tracking
wandb>=0.16.0

# HuggingFace Hub
huggingface_hub>=0.20.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0
rich>=13.7.0
